# PWX-43070: fix race condition in fastpath failover