** UT notes **

To run UT, define `TEST_FP_RACE` to 1 in `pxd.h`,
 

 
 ** Before fix - IOs gets stuck **
 ```
 ./t --gtest_filter=*FastpathRace*
Note: Google Test filter = *FastpathRace*
[==========] Running 1 test from 1 test case.
[----------] Global test environment set-up.
[----------] 1 test from FailoverTest
[ RUN      ] FailoverTest.FastpathRace
25600+0 records in
25600+0 records out
104857600 bytes (105 MB, 100 MiB) copied, 0.397286 s, 264 MB/s
mke2fs 1.44.1 (24-Mar-2018)
Discarding device blocks: done                            
Creating filesystem with 25600 4k blocks and 25600 inodes

Allocating group tables: done                            
Writing inode tables: done                            
Creating journal (1024 blocks): done
Writing superblocks and filesystem accounting information: done

time="2025-04-10 18:07:39Z" level=INFO msg="Executing: sudo dmsetup create flakey_test_loop10 < /tmp/dm_table_sJvGdz"
time="2025-04-10 18:07:39Z" level=INFO msg="Executing: sudo dmsetup create delay_test_loop10 < /tmp/dm_table_IjNEqx"
time="2025-04-10 18:07:39Z" level=INFO msg="Added device /dev/mapper/delay_test_loop10 minor 1 dev_id 1234"
time="2025-04-10 18:07:39Z" level=INFO msg="Pxd path /dev/pxd/pxd1234"
time="2025-04-10 18:07:39Z" level=INFO msg="[fio] executing: sudo fio --name=test --filename=/dev/pxd/pxd1234 --ioengine=libaio --rw=write --bs=4k --offset=52428800 --size=4k --direct=1 --iodepth=256 --numjobs=16"
test: (g=0): rw=write, bs=(R) 4096B-4096B, (W) 4096B-4096B, (T) 4096B-4096B, ioengine=libaio, iodepth=256
...
fio-3.1
Starting 16 processes
time="2025-04-10 18:07:46Z" level=INFO msg="[fio] executing: sudo fio --name=test --filename=/dev/pxd/pxd1234 --ioengine=libaio --rw=write --bs=4k --offset=52428800 --size=4k --direct=1 --iodepth=256 --numjobs=16"
test: (g=0): rw=write, bs=(R) 4096B-4096B, (W) 4096B-4096B, (T) 4096B-4096B, ioengine=libaio, iodepth=256
...
fio-3.1
Starting 16 processes
time="2025-04-10 18:07:47Z" level=INFO msg="wb = 24"0,w=0 IOPS][eta 00m:00s]
time="2025-04-10 18:07:47Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:07:47Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:07:47Z" level=INFO msg="Processing request opcode: 8208 minor: 1 num_volumes: 1"
time="2025-04-10 18:07:47Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:07:47Z" level=INFO msg="Got failover request"
Jobs: 16 (f=16): [W(16)][-.-%][r=0KiB/s,w=0KiB/s][r=0,w=0 IOPS][eta 00m:00s]
Jobs: 16 (f=16): [W(16)][-.-%][r=0KiB/s,w=0KiB/s][r=0,w=0 IOPS][eta 00m:00s]
Jobs: 16 (f=16): [W(16)][-.-%][r=0KiB/s,w=0KiB/s][r=0,w=0 IOPS][eta 00m:00s]
Jobs: 16 (f=16): [W(16)][-.-%][r=0KiB/s,w=0KiB/s][r=0,w=0 IOPS][eta 00m:00s]
Jobs: 16 (f=16): [W(16)][-.-%][r=0KiB/s,w=0KiB/s][r=0,w=0 IOPS][eta 00m:00s]
Jobs: 16 (f=16): [W(16)][-.-%][r=0KiB/s,w=0KiB/s][r=0,w=0 IOPS][eta 00m:00s]
Jobs: 16 (f=16): [W(16)][-.-%][r=0KiB/s,w=0KiB/s][r=0,w=0 IOPS][eta 00m:00s]
Jobs: 16 (f=16): [W(16)][-.-%][r=0KiB/s,w=0KiB/s][r=0,w=0 IOPS][eta 00m:00s]
Jobs: 16 (f=16): [W(16)][-.-%][r=0KiB/s,w=0KiB/s][r=0,w=0 IOPS][eta 00m:00s]
Jobs: 16 (f=16): [W(16)][-.-%][r=0KiB/s,w=0KiB/s][r=0,w=0 IOPS][eta 00m:00s]
```
 

 ** After fix **
 
 ```
 ./t --gtest_filter=*FastpathRace*
Note: Google Test filter = *FastpathRace*
[==========] Running 1 test from 1 test case.
[----------] Global test environment set-up.
[----------] 1 test from FailoverTest
[ RUN      ] FailoverTest.FastpathRace
25600+0 records in
25600+0 records out
104857600 bytes (105 MB, 100 MiB) copied, 0.421346 s, 249 MB/s
mke2fs 1.44.1 (24-Mar-2018)
Discarding device blocks: done                            
Creating filesystem with 25600 4k blocks and 25600 inodes

Allocating group tables: done                            
Writing inode tables: done                            
Creating journal (1024 blocks): done
Writing superblocks and filesystem accounting information: done

time="2025-04-10 18:01:06Z" level=INFO msg="Executing: sudo dmsetup create flakey_test_loop10 < /tmp/dm_table_X78f7W"
time="2025-04-10 18:01:06Z" level=INFO msg="Executing: sudo dmsetup create delay_test_loop10 < /tmp/dm_table_9SOnTU"
time="2025-04-10 18:01:06Z" level=INFO msg="Added device /dev/mapper/delay_test_loop10 minor 1 dev_id 1234"
time="2025-04-10 18:01:06Z" level=INFO msg="Pxd path /dev/pxd/pxd1234"
time="2025-04-10 18:01:06Z" level=INFO msg="[fio] executing: sudo fio --name=test --filename=/dev/pxd/pxd1234 --ioengine=libaio --rw=write --bs=4k --offset=52428800 --size=4k --direct=1 --iodepth=256 --numjobs=16"
test: (g=0): rw=write, bs=(R) 4096B-4096B, (W) 4096B-4096B, (T) 4096B-4096B, ioengine=libaio, iodepth=256
...
fio-3.1
Starting 16 processes
time="2025-04-10 18:01:13Z" level=INFO msg="[fio] executing: sudo fio --name=test --filename=/dev/pxd/pxd1234 --ioengine=libaio --rw=write --bs=4k --offset=52428800 --size=4k --direct=1 --iodepth=256 --numjobs=16"
test: (g=0): rw=write, bs=(R) 4096B-4096B, (W) 4096B-4096B, (T) 4096B-4096B, ioengine=libaio, iodepth=256
...
fio-3.1
Starting 16 processes
time="2025-04-10 18:01:14Z" level=INFO msg="wb = 24"0,w=0 IOPS][eta 00m:00s]
time="2025-04-10 18:01:14Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:14Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:14Z" level=INFO msg="Processing request opcode: 8208 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:14Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:14Z" level=INFO msg="Got failover request"
time="2025-04-10 18:01:19Z" level=INFO msg="Added device /dev/mapper/flakey_test_loop10 minor 1 dev_id 1234"
time="2025-04-10 18:01:19Z" level=INFO msg="Pxd path /dev/pxd/pxd1234"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8208 unique: 262144"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 17"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262272"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 16"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262273"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 15"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262274"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 14"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262275"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 13"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262276"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 12"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262400"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 11"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262528"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 10"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262656"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 9"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262401"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 8"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262529"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 7"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262402"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 6"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262657"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 5"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262530"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 4"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262531"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 3"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262403"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8193 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got write request - offset: 52428800 length: 4096 active_io_count 2"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8193 unique: 262658"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 104849408 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262277"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 0 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262278"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 4096 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262279"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 104853504 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262280"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 104722432 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262281"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 104824832 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262282"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 104726528 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262283"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 104652800 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262284"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 104554496 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262285"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 104509440 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262286"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 104480768 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262287"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 104390656 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262288"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 104357888 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262289"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 104349696 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262290"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 104370176 length: 4096"
time="2025-04-10 18:01:19Z" level=INFO msg="sending reply for opcode : 8194 unique: 262291"
time="2025-04-10 18:01:19Z" level=INFO msg="epoll_wait returned 1"
time="2025-04-10 18:01:19Z" level=INFO msg="got event from kernel"
time="2025-04-10 18:01:19Z" level=INFO msg="Processing request opcode: 8194 minor: 1 num_volumes: 1"
time="2025-04-10 18:01:19Z" level=INFO msg="active_idx : 0"
time="2025-04-10 18:01:19Z" level=INFO msg="Got read request - offset: 103276544 length: 4096"
…