@sreymondjohn 2025-04-11
I had an idea to solve this via rcu. Let me check that, for now closing the PR, will re-open if the rcu one doesn't work.